src/main.cpp: store hsm_so_pin in fiovb environment
This patch is part of a set of patches across multiple repositories in
order to add support from cleanup of the RPMB in case of a full factory
reset.

When a hsm config has been provided, after the `sota.toml` has been
written also store the hsm_so_pin which was used to initialize the token
in the fiovb environment.

Signed-off-by: Bram Vlerick <bram.vlerick@openpixelsystems.org>